backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymongo
passlib
pandas
numpy
reportlab
python-multipart
